feat: add app service to docker-compose.yaml
An 'app' service was added to the docker-compose.yaml file. This service uses the 'enhanced-preprints-client' image from the eLife Sciences repository. A healthcheck with a curl command was implemented. Moreover, it depends on the 'api' service being healthy and maps the port 3001 to the container's port 3000. Several environment variables were also set.