Add test for revised preprints (#27)
* Create initial structure for tests

* Rename test folders/files for more clarity

* Add initial mock data for wiremock

* feat: add next-step actions, assertions and inputs in dummy-1.json

This commit includes the addition of next-step actions, assertions, and
inputs in the dummy-1.json file. It provides a detailed structure for
the next step which includes preprint outputs, under-review and draft
assertions, and preprint inputs.

* feat: add next-step actions, assertions and inputs in dummy-1.json

This commit includes the addition of next-step actions, assertions, and
inputs in the dummy-1.json file. It provides a detailed structure for
the next step which includes preprint outputs, under-review and draft
assertions, and preprint inputs.

* feat: update reviews.spec.ts and revise.spec.ts

Update reviews.spec.ts to dynamically delete preprints instead of hardcoding the name.
Add a new test to revised.spec.ts to verify the availability of revised preprints.

* feat: add assertions for article status in revised.spec.ts test

This commit includes more specific assertions in the revised.spec.ts
test. It now checks that the article status text correctly reflects
'Published from the original preprint after peer review and
assessment by eLife.' and 'Revised by authors after peer review.'

* Add assertions for the doi version on each version page, and navigating between versions. Also check that the msid URL loads the second version

* Add assertions for correctly displaying "(this version)" in the timeline

* Add a "this version" check for version 1 as well

* feat: Improve readability and add timeline checks in revised test

This commit updates the revised preprint test to improve readability and
add further checks on the review timeline. It removes excessive blank
lines and corrects spelling mistakes in the comments. It also adds
additional check for the latest revision in the review timeline.

---------

Co-authored-by: aisha <a.kapan@elifesciences.org>
Co-authored-by: Scott Aubrey <scott@aubrey.org.uk>