Introduce preview and publication of revised preprint to progress test (#28)
* feat: add detailed actions and assertions to dummy-1-published.json

This commit adds a host of actions, outputs, participants, and assertions
to the `dummy-1-published.json` file. This enhancement provides more
comprehensive mock data for testing purposes.

* Create initial structure for tests

* Rename test folders/files for more clarity

* Add initial mock data for wiremock

* feat: Add mock data and wiremock mappings for progress scenario

This commit introduces mock data for progress scenario in the
dummy-1-published-and-preview.json file located under the
mock-data/docmaps/progress directory. In addition, it also includes wiremock
mappings related to the same scenario in
dummy-1-published-and-preview.json and index-published-and-preview.json files
under the wiremock/mappings/progress directory.

* feat: add next-step actions, assertions and inputs in dummy-1.json

This commit includes the addition of next-step actions, assertions, and
inputs in the dummy-1.json file. It provides a detailed structure for
the next step which includes preprint outputs, under-review and draft
assertions, and preprint inputs.

* feat: add next-step actions, assertions and inputs in dummy-1.json

This commit includes the addition of next-step actions, assertions, and
inputs in the dummy-1.json file. It provides a detailed structure for
the next step which includes preprint outputs, under-review and draft
assertions, and preprint inputs.

* feat: update reviews.spec.ts and revise.spec.ts

Update reviews.spec.ts to dynamically delete preprints instead of hardcoding the name.
Add a new test to revised.spec.ts to verify the availability of revised preprints.

* feat: add assertions for article status in revised.spec.ts test

This commit includes more specific assertions in the revised.spec.ts
test. It now checks that the article status text correctly reflects
'Published from the original preprint after peer review and
assessment by eLife.' and 'Revised by authors after peer review.'

* Add assertions for the doi version on each version page, and navigating between versions. Also check that the msid URL loads the second version

* Add assertions for correctly displaying "(this version)" in the timeline

* Add a "this version" check for version 1 as well

* feat: Improve readability and add timeline checks in revised test

This commit updates the revised preprint test to improve readability and
add further checks on the review timeline. It removes excessive blank
lines and corrects spelling mistakes in the comments. It also adds
additional check for the latest revision in the review timeline.

* Refactor progress test and fixtures to introduce revised preprint preview and published

* Address formatting

* Amend revised test with different meca for revised preprint

* feat: update status code check in progress.spec.ts

Changed the expected status code from 404 to any status code greater than 400 in the reviewed-preprints test.

* feat: update manuscript identifiers in mock data

Replace 'revised-msid' with 'progress-msid' in dummy-1-preview-revised.json and dummy-1-revised.json files in mock data. The changes affects various properties including 'outputs' and 'partOf'.

* feat: Increase timeout in playwright config

The timeout value in the playwright configuration file has been
increased from 240000 to 360000 to allow for longer test execution.

* feat: update test assertions for title text in revised.spec.ts

This commit updates assertions in the revised.spec.ts test file to
use .toHaveText instead of .toContainText for 'h1.title' locator. This
enhances specificity in validating the exact title text.

* feat: update test assertions for title text in revised.spec.ts

This commit updates assertions in the revised.spec.ts test file to
use .toHaveText instead of .toContainText for 'h1.title' locator. This
enhances specificity in validating the exact title text.

---------

Co-authored-by: aisha <a.kapan@elifesciences.org>
Co-authored-by: Scott Aubrey <scott@aubrey.org.uk>