feat: add test for import continuation on approval signal
This commit includes a new test for import continuation when the approval
signal is set to true. The test ensures that the workflow status changes
to 'COMPLETED' and the response from 'awaitingApproval' query is null.
Additionally, it checks the success status of axios GET requests for
'preprints/{name}-msidv1' and 'preprints/{name}-msid-2v1' endpoints. The
commit also includes a cleanup step to delete S3 Epp folders and axios
DELETE requests for the same endpoints.